renamed the simple_network.py -> resnet_18.py. And the resnet18 link to ReadMe.
