+++ examples/hello-world/hello-pt-resnet/README.md
@@ -10,7 +10,9 @@ instead of the SimpleNetwork.
 The Job API only supports the object instance created directly out of the Python Class. It does not support 
 the object instance created through using the Python function. Comparing with the hello-pt example, 
 if we replace the SimpleNetwork() object with the resnet18(num_classes=10), 
-the "resnet18(num_classes=10)" creates an torchvision "ResNet" object instance out of the "resnet18" function. The job API can 
+the "resnet18(num_classes=10)" creates an torchvision "ResNet" object instance out of the "resnet18" function. 
+As shown in the [torchvision reset](https://github.com/pytorch/vision/blob/main/torchvision/models/resnet.py#L684-L705), 
+the resnet18 is a Python function, which creates and returns a ResNet object. The job API can 
 only use the "ResNet" object instance for generating the job config. It can not detect the object creating function logic in the "resnet18".
 
 This example demonstrates how to wrap up the resnet18 Python function into a Resnet18 Python class. Then uses the Resnet18(num_classes=10)
